requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg[binary]>=3.0.0
python-dotenv==0.19.0
openai==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
email-validator==1.1.3
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.19.0 